+++ pyengine/Entity.py
@@ -78,19 +78,6 @@ class Entity(pygame.sprite.Sprite):
         if self.has_component(PhysicsComponent):
             self.get_component(PhysicsComponent).update_gravity()
 
-        if self.has_component(PositionComponent):
-            from pyengine import WindowCallbacks
-            # Verify if entity is not out of window
-            position = self.get_component(PositionComponent)
-            if position.position.y >= self.system.world.window.height - self.image.get_rect().height:
-                self.system.world.window.call(WindowCallbacks.OUTOFWINDOW, self, position.position)
-            elif position.position.y < 0:
-                self.system.world.window.call(WindowCallbacks.OUTOFWINDOW, self, position.position)
-            if position.position.x >= self.system.world.window.width - self.image.get_rect().width:
-                self.system.world.window.call(WindowCallbacks.OUTOFWINDOW, self, position.position)
-            elif position.position.x < 0:
-                self.system.world.window.call(WindowCallbacks.OUTOFWINDOW, self, position.position)
-
         if self.has_component(ControlComponent):
             self.get_component(ControlComponent).update()